# server/openapi3-spec.yaml
openapi: 3.0.0
paths:
...
  /users:
    post:
      operationId: UsersController_createUser
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserDto'
      responses:
        '201':
          description: ''
    get:
      operationId: UsersController_listUsers
      parameters: []
      responses:
        '200':
          description: ''
  /users/{id}:
    get:
      operationId: UsersController_getUser
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: ''
info:
  title: API Gateway + Cloud Run
  description: Sample API on APIGateway with a Cloud Run backend
  version: 1.0.0
  contact: {}
...
servers: []
components:
  schemas:
...
    CreateUserDto:
      type: object
      properties: {}